dom/jsurl/test/test_bug351633-1.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=351633
-->
<head>
  <title>Test for Bug 351633: Form submission</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=351633">Mozilla Bug 351633</a>
<p id="display">
  <iframe id="frame" src="form-submit.html"></iframe>
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 351633 **/
var passJSUrl = false;

function finishTest() {
  is(passJSUrl, true, "Unexpected result");
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();

function runTest() {
  $("frame").contentDocument.getElementById("test").submit();
}

addLoadEvent(runTest);

</script>
</pre>
</body>
</html>
